Fix: Too many files included
When there is a subdir which fits within upload_limit,
it was previously backed up in full, even if only a subset
of the contents was listed in backup_paths.

This stems from the time when only first-level dirs
of the pool could be specified.

Rewrite the set building to fix this and add tests.